+++ AlchemistCode/Cards/Common/Spike.cs
@@ -17,6 +17,6 @@ public class Spike : AlchemistCard
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_attack_slash")).Execute(choiceContext);
-        await Mixing.CreateOne<Token.FumingMix>(choiceContext, Owner, IsUpgraded);
+        await Mixing.CreateOne<Token.FumingMix>(choiceContext, Owner, IsUpgraded, this);
     }
 }